REST API
Web Requests and Bots in JavaScript
Created a program where I can send a message to a specific channel and program would respond with the weather data in specific city, implemented using REST API and Web Requests, where JSON (JavaScript Object Notation) and GSON are heavily used, also created a simple program that can connect to a Chat Server. The ChatServer used was a FreeNode. I exposed my bot so anyone can interact with it, When the bot is connected to the group chat it could read the messages and when a message matches a command it respond accordingly, if someone says 􏰀Weather 75080􏰁 for example it returns the weather API and return the information in a IRC-friendly format.